requirements.txt
mcp
fastmcp
langchain-mcp-adapters
langchain-google-genai
langgraph
python-dotenv